Migrate all user and note IDs to snowflakes

+++ cmd/internal/domain/entity/connection.go
@@ -12,7 +12,7 @@ const (
 
 type Connection struct {
 	ConnectionID    string `gorm:"primaryKey;autoIncrement:false"`
-	UserID          int    `gorm:"not null;index"`
+	UserID          int64  `gorm:"not null;index"`
 	ExpiresAt       int64  `gorm:"not null"`
 	LastHeartbeatAt int64  `gorm:"not null;index"`
 	CreatedAt       int64  `gorm:"not null"`
